Change default catalog from hive to system
Currently, the default catalog for creating connection string is set to
hive. This could be a problem if no catalog is specified in the connection
URL and then hive catalog is used even if it is not configured on the Trino
side.This leads to a connection error on Superset UI. To avoid this it is
better to use generic catalog such as 'system'.

Closes #13